+++ changelog.md
@@ -5,3 +5,8 @@
 - Add changelog.md
 - Add READEME.md
 - Initiate HTML, CSS, JS and Folders
+
+**v1.0.0**
+- It is now completed
+- Completed after 1 hour of coding
+- Added Images assets (Credits for those images)